Remove pronatec, edit import export serializers
